Renamed Workout ui model. Added init method in workout presenter.
