+++ app/src/main/kotlin/com/dvinc/circlestimer/presentation/ui/workout/WorkoutPresenter.kt
@@ -5,15 +5,18 @@
 
 package com.dvinc.circlestimer.presentation.ui.workout
 
-import com.dvinc.circlestimer.presentation.model.workout.Workout
+import com.dvinc.circlestimer.domain.model.workout.Workout
+import com.dvinc.circlestimer.domain.usecase.workout.WorkoutUseCase
 import com.dvinc.circlestimer.presentation.ui.base.BasePresenter
+import javax.inject.Inject
 
-class WorkoutPresenter : BasePresenter<WorkoutView>() {
+class WorkoutPresenter @Inject constructor(
+        private val workoutUseCase: WorkoutUseCase
+) : BasePresenter<WorkoutView>() {
 
-    //TODO: Inject interactor
-
-    fun initWorkoutView() {
-        //TODO: load workouts here
+    override fun attachView(view: WorkoutView) {
+        super.attachView(view)
+        initWorkoutView()
     }
 
     fun onItemSwiped(workout: Workout)  {
@@ -22,5 +25,14 @@ class WorkoutPresenter : BasePresenter<WorkoutView>() {
 
     fun onWorkoutDeleted(workout: Workout) {
         //TODO: delete workout. And rename this method?
+    }
+
+    //TODO: handle error
+    private fun initWorkoutView() {
+        addSubscription(workoutUseCase.obtainWorkouts()
+                .subscribe(
+                        { getView()?.showWorkouts(it) },
+                        { it.printStackTrace() }
+                ))
     }
 }
